chore: consolidate StorageITParamRunner and StorageSuite into StorageITRunner
